conversation already working, but have to implement filters

+++ src/main/java/wextechnicaltestLeandroSenaZuza/wexpurchaseapi/model/TransactionPurchase.java
@@ -2,8 +2,11 @@ package wextechnicaltestLeandroSenaZuza.wexpurchaseapi.model;
 
 import jakarta.persistence.*;
 import lombok.Data;
+import org.springframework.format.annotation.DateTimeFormat;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
+import java.text.SimpleDateFormat;
 import java.util.Date;
 
 @Data
@@ -17,11 +20,11 @@ public class TransactionPurchase {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "CODG_TRANSACTION")
     private Long idTransaction;
-    @Temporal(TemporalType.TIMESTAMP)
+    @Temporal(TemporalType.TIMESTAMP )
     @Column(name = "DATE_TRANSACTION", nullable = false)
     private Date transactionDate;
     @Column(name = "PURCHASE_AMOUNT", nullable = false)
-    private double purchaseAmount;
+    private BigDecimal purchaseAmount;
     @Column(name = "DESCRIPTION", length = 50, nullable = true)
     private String description;
 }
